requirements.txt
# Web
fastapi>=0.111,<1
uvicorn[standard]>=0.30,<1
python-multipart>=0.0.9
pydantic>=2.8,<3
python-dotenv>=1.0,<2

# DOCX only
python-docx>=1.1,<2

# Cognee (pin the version you tested)
cognee==0.3.6

# Keep NumPy on 1.x because some transitive wheels still target it
numpy==1.26.4

dotenv==1.1.1
openai
